requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
cryptography==3.2.1
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
gunicorn==20.0.4
jwt==1.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
